+++ src/util/util.shared.ts
@@ -1,9 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import { BigNumber } from 'bignumber.js'
 import { Container } from 'typedi'
 import { version } from '../../package.json'
-import { BigNumber } from 'bignumber.js'
 import {
-  AddressEventNotificationChain, isEvmArchiveNonArchiveLoadBalancerNetwork,
+  AddressEventNotificationChain,
+  isEvmArchiveNonArchiveLoadBalancerNetwork,
   isEvmBasedNetwork,
   isEvmLoadBalancerNetwork,
   isSolanaEnabledNetwork,
@@ -46,7 +47,8 @@ import {
   Palm,
   Polygon,
   Solana,
-  SolanaRpc, TatumConfig,
+  SolanaRpc,
+  TatumConfig,
   Tron,
   TronRpc,
   UtxoLoadBalancerRpc,
@@ -56,8 +58,8 @@ import {
   Xrp,
   XrpRpc,
 } from '../service'
-import { CONFIG } from './di.tokens'
 import { EvmArchiveLoadBalancerRpc } from '../service/rpc/evm/EvmArchiveLoadBalancerRpc'
+import { CONFIG } from './di.tokens'
 
 export const Utils = {
   getRpc: <T>(id: string, config: TatumConfig): T => {
@@ -215,14 +217,14 @@ export const Utils = {
     clearTimeout(id)
     return { responseTime, response }
   },
-  getHeaders: (id: string, retry = 0) => {
+
+  getHeaders: (id: string) => {
     const config = Container.of(id).get(CONFIG)
     const headers = new Headers({
       'Content-Type': 'application/json',
       'x-ttm-sdk-version': version,
       'x-ttm-sdk-product': 'JS',
       'x-ttm-sdk-debug': `${config.verbose}`,
-      'x-ttm-sdk-retry': `${retry}`,
     })
 
     if (config.apiKey) {